+++ test/canvas.test.js
@@ -1623,22 +1623,41 @@ describe('Canvas', function () {
     var canvas = createCanvas(500, 500);
     var ctx = canvas.getContext('2d');
 
+    // Drawing canvas to itself
     ctx.fillStyle = 'white';
     ctx.fillRect(0, 0, 500, 500);
     ctx.fillStyle = 'black';
     ctx.fillRect(5, 5, 10, 10);
-    ctx.drawImage(ctx.canvas, 20, 20);
+    ctx.drawImage(canvas, 20, 20);
 
     var imgd = ctx.getImageData(0, 0, 500, 500);
     var data = imgd.data;
     var count = 0;
 
-    for(var i = 0; i < 500 * 500; i += 4){
+    for(var i = 0; i < 500 * 500 * 4; i += 4){
       if(data[i] === 0 && data[i + 1] === 0 && data[i + 2] === 0)
         count++;
     }
 
     assert.strictEqual(count, 10 * 10 * 2);
+
+    // Drawing zero-width image
+    ctx.drawImage(canvas, 0, 0, 0, 0, 0, 0, 0, 0);
+    ctx.drawImage(canvas, 0, 0, 0, 0, 1, 1, 1, 1);
+    ctx.drawImage(canvas, 1, 1, 1, 1, 0, 0, 0, 0);
+    ctx.fillStyle = 'white';
+    ctx.fillRect(0, 0, 500, 500);
+
+    imgd = ctx.getImageData(0, 0, 500, 500);
+    data = imgd.data;
+    count = 0;
+
+    for(i = 0; i < 500 * 500 * 4; i += 4){
+      if(data[i] === 255 && data[i + 1] === 255 && data[i + 2] === 255)
+        count++;
+    }
+
+    assert.strictEqual(count, 500 * 500);
   });
 
   it('Context2d#SetFillColor()', function () {